+++ include/helpers/measurements.hpp
@@ -12,11 +12,11 @@ public:
     double initialization_time;
     std::vector<double> adaptation_time;
     std::vector<double> query_time;
-    std::vector<size_t> max_height;
-    std::vector<size_t> min_height;
-    std::vector<size_t> number_of_nodes;
-    std::vector<size_t> memory_footprint;
-    std::vector<size_t> tuples_scanned;
+    std::vector<int64_t> max_height;
+    std::vector<int64_t> min_height;
+    std::vector<int64_t> number_of_nodes;
+    std::vector<int64_t> memory_footprint;
+    std::vector<int64_t> tuples_scanned;
 
     Measurements();
     ~Measurements();
