+++ include/indexes/quasii.hpp
@@ -29,20 +29,20 @@ private:
     vector<Slice> first_level_slices;
 
     int64_t last_level_threshold = 2000;
-    vector<size_t> dimensions_threshold;
+    vector<int64_t> dimensions_threshold;
 
-    size_t count_slices(vector<Slice> &slices);
+    int64_t count_slices(vector<Slice> &slices);
 
     void adapt(Query& query);
 
-    vector<pair<size_t, size_t> > search(Query& query);
-    size_t binarySearch(const vector<Slice> &slice, float key);
+    vector<pair<int64_t, int64_t> > search(Query& query);
+    int64_t binarySearch(const vector<Slice> &slice, float key);
 
-    size_t predicate_index(size_t column, Query &query);
+    int64_t predicate_index(int64_t column, Query &query);
 
-    bool column_in_query(size_t column, Query &query);
+    bool column_in_query(int64_t column, Query &query);
 
-    Predicate predicate_on_column(size_t column, Query& query);
+    Predicate predicate_on_column(int64_t column, Query& query);
 
     void calculate_level_thresholds();
